src/requirements.txt
dgl==2.1.0
fgvcdata==0.1.0
hydra-core==1.3.2
imageio==2.33.1
matplotlib==3.3.4
networkx==2.5.1
nni==2.6.1
numba==0.53.1
numpy==1.19.2
omegaconf==2.3.0
opencv_python==4.2.0.32
pandas==1.1.5
Pillow==10.2.0
pytorch_lightning==1.5.10
PyYAML==6.0.1
scikit_learn==1.3.2
scipy==1.5.2
seaborn==0.11.2
six==1.16.0
timm==0.6.12
torch==1.13.1
torchmetrics==0.11.2
torchvision==0.14.1
tqdm==4.64.1
wandb==0.16.2